+++ ci_cd/remote/host/variables.tf
@@ -1,19 +1,6 @@
-variable "connection_map" {
-  type = map(object({
-    host_key      = optional(string)
-    provider_type = optional(string)
-  }))
-  description = "These are created in pending state and authorized by whomever does it. Therefore they are linked to the permissions for a user."
-}
-
-variable "connection_provider_type_default" {
-  type    = string
-  default = "GitHub"
-}
-
 variable "host_map" {
   type = map(object({
-    name_override               = optional(string)
+    name_override               = optional(string) # This is for imported hosts
     provider_endpoint           = string
     provider_type               = optional(string)
     vpc_az_key_list             = optional(list(string))
@@ -35,38 +22,12 @@ variable "host_vpc_tls_certificate_default" {
   default = null
 }
 
-variable "policy_create" {
-  type    = bool
-  default = true
-}
-
-variable "policy_name" {
-  type        = string
-  default     = null
-  description = "Defaults to code-connection"
-}
-
-variable "policy_name_infix" {
-  type    = bool
-  default = true
-}
-
-variable "policy_name_prefix" {
-  type    = string
-  default = ""
-}
-
 variable "std_map" {
   type = object({
-    access_title_map               = map(string)
-    aws_account_id                 = string
-    aws_region_name                = string
-    iam_partition                  = string
-    name_replace_regex             = string
-    resource_name_prefix           = string
-    resource_name_suffix           = string
-    service_resource_access_action = map(map(map(list(string))))
-    tags                           = map(string)
+    name_replace_regex   = string
+    resource_name_prefix = string
+    resource_name_suffix = string
+    tags                 = map(string)
   })
 }