+++ iam/app/lambda/main.tf
@@ -1,0 +1,5 @@
+module "lambda_net_policy" {
+  source  = "../../../iam/policy/identity/lambda/net"
+  name    = "lambda_vpc"
+  std_map = var.std_map
+}