+++ iam/policy/identity/cdn/distribution/main.tf
@@ -1,5 +1,5 @@
 module "this_policy" {
-  source      = "../../../../iam/policy/identity/access_resource"
+  source      = "../../../../../iam/policy/identity/access_resource"
   access_list = var.access_list
   name        = var.name
   name_infix  = var.name_infix