+++ lambda/function/locals.tf
@@ -6,7 +6,7 @@ module "name_map" {
 }
 
 module "policy_map" {
-  source                      = "../../iam/policy_name_map"
+  source                      = "../../iam/policy/name_map"
   name_map                    = var.function_map
   policy_access_list_default  = var.policy_access_list_default
   policy_create_default       = var.policy_create_default