+++ secret/random/outputs.tf
@@ -1,3 +1,10 @@
 output "data" {
   value = local.output_data
 }
+
+output "secret_map" {
+  value = merge(
+    module.this_param.secret_map,
+    module.this_secret.secret_map
+  )
+}