+++ ssm/parameter_secret/variables.tf
@@ -1,23 +1,47 @@
 variable "param_map" {
   type = map(object({
-    kms_key_id             = optional(string)
-    policy_access_list     = optional(list(string))
-    policy_create          = optional(bool)
-    policy_name            = optional(string)
-    policy_name_append     = optional(string)
-    policy_name_infix      = optional(bool)
-    policy_name_prefix     = optional(string)
-    policy_name_prepend    = optional(string)
-    policy_name_suffix     = optional(string)
-    secret_random_init_key = optional(string)
-    secret_random_init_map = optional(map(string))
-    tier                   = optional(string)
+    kms_key_id              = optional(string)
+    name_include_app_fields = optional(bool)
+    name_infix              = optional(bool)
+    name_prefix             = optional(string)
+    name_suffix             = optional(string)
+    policy_access_list      = optional(list(string))
+    policy_create           = optional(bool)
+    policy_name             = optional(string)
+    policy_name_append      = optional(string)
+    policy_name_infix       = optional(bool)
+    policy_name_prefix      = optional(string)
+    policy_name_prepend     = optional(string)
+    policy_name_suffix      = optional(string)
+    secret_random_init_key  = optional(string)
+    secret_random_init_map  = optional(map(string))
+    tier                    = optional(string)
   }))
 }
 
 variable "param_kms_key_id_default" {
   type    = string
   default = "alias/aws/ssm"
+}
+
+variable "param_name_include_app_fields_default" {
+  type    = bool
+  default = true
+}
+
+variable "param_name_infix_default" {
+  type    = bool
+  default = true
+}
+
+variable "param_name_prefix_default" {
+  type    = string
+  default = ""
+}
+
+variable "param_name_suffix_default" {
+  type    = string
+  default = ""
 }
 
 variable "param_secret_random_init_key_default" {